Realized sending all Kademlia responses

+++ DiscoveryService/Messages/Message.cs
@@ -84,7 +84,7 @@ namespace LUC.DiscoveryService.Messages
         {
             if(reader != null)
             {
-                MessageOperation = (MessageOperation)reader.ReadByte();
+                MessageOperation = (MessageOperation)reader.ReadUInt32();
                 return this;
             }
             else
